Make shared functionality for shader cursor
